我也不知道更新了啥

+++ 第一次作业/6计算分数值.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 void printSum()
 {
-    double res = 0;
+    float res = 0;
     for (int i = 1; i <= 100; i++)
     {
         if (i % 2)
